requirements.txt
# Core dependencies
fastapi==0.103.0
uvicorn==0.22.0
pymongo[srv]==4.2.0
pydantic==2.3.0
python-dotenv==1.0.0

# Development dependencies
pytest==7.2.2
pytest-asyncio==0.21.0
pytest-cov==3.0.0
